requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py38hca72f7f_0
ca-certificates=2022.07.19=hecd8cb5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
gst-plugins-base=1.14.0=h4180768_2
importlib-metadata=4.11.3=py38hecd8cb5_0
intel-openmp=2021.4.0=hecd8cb5_3538
libwebp-base=1.2.2=hca72f7f_0
llvm-openmp=12.0.0=h0dcd299_1
lz4-c=1.9.3=h23ab428_1
matplotlib-base=3.5.1=py38hfb0c5b7_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mkl-service=2.4.0=py38h9ed2024_0
nest-asyncio=1.5.5=py38hecd8cb5_0
numpy-base=1.23.1=py38h3b1a694_0
opt-einsum=3.3.0=pypi_0
paddle-bfloat=0.1.7=pypi_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
pyqt5-sip=12.11.0=py38he9d5cce_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
python-louvain=0.15=pyhd3eb1b0_0
pytorch-cluster=1.6.0=py38_torch_1.11.0_cpu
pytorch-scatter=2.0.9=py38_torch_1.11.0_cpu
pytorch-sparse=0.6.14=py38_torch_1.11.0_cpu
pytorch-spline-conv=1.2.1=py38_torch_1.11.0_cpu
qt-main=5.15.2=h719ae48_4
qt-webengine=5.15.9=h90a370e_4
scikit-learn=1.1.1=py38he9d5cce_0
typing-extensions=4.1.1=hd3eb1b0_0
websocket-client=0.58.0=py38hecd8cb5_4
